Fix crash in Qt3To5UIClasses
It appears that asking any CXXRecordDecl for its bases crashes should
that decl not have any bases. There doesn't appear to be a way to check
for it ahead of time without crashing, so I followed the lead of some
clang source that checked isThisDeclarationADefinition before calling
bases. I'm not sure it is the proper fix but it appears to resolve the
crashing and work on the example.